+++ tiled/server/pydantic_sparse.py
@@ -2,7 +2,7 @@ from typing import Optional, Tuple, Union
 
 import pydantic
 
-from ..structures.array import BuiltinDtype, StructDtype
+from ..structures.array import BuiltinDtype, StructDtype, Endianness, Kind
 from ..structures.sparse import SparseLayout
 
 
@@ -10,6 +10,7 @@ class COOStructure(pydantic.BaseModel):
     shape: Tuple[int, ...]  # tuple of ints like (3, 3)
     chunks: Tuple[Tuple[int, ...], ...]  # tuple-of-tuples-of-ints like ((3,), (3,))
     data_type: Optional[Union[BuiltinDtype, StructDtype]] = None
+    indx_data_type: BuiltinDtype = BuiltinDtype(Endianness("little"), Kind("u"), 8)   # numpy 'uint' dtype
     dims: Optional[Tuple[str, ...]] = None  # None or tuple of names like ("x", "y")
     resizable: Union[bool, Tuple[bool, ...]] = False
     layout: SparseLayout = SparseLayout.COO